refactor(extra-info): extract extra info buttons to a separate component
